._old/cloud_comparison/requirements.txt
boto3>=1.26.0
azure-ai-videoanalyzer>=1.0.0
azure-identity>=1.12.0
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
python-dotenv>=1.0.0 